프로젝트 파일 구조 변경 - doamin     - image     - interestfield     - keyword     - qna     - searchlog     - tag     - users     - wish
