+++ src/main/java/com/strawberryfarm/fitingle/TestController/Member.java
@@ -1,0 +1,11 @@
+package com.strawberryfarm.fitingle.TestController;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+public class Member {
+	@Id
+	private int id;
+}
