+++ src/main/java/com/strawberryfarm/fitingle/domain/AlertLog.java
@@ -1,5 +1,6 @@
 package com.strawberryfarm.fitingle.domain;
 
+import com.strawberryfarm.fitingle.domain.users.entity.Users;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
