+++ src/main/java/com/strawberryfarm/fitingle/domain/Board.java
@@ -1,5 +1,9 @@
 package com.strawberryfarm.fitingle.domain;
 
+import com.strawberryfarm.fitingle.domain.image.entity.Image;
+import com.strawberryfarm.fitingle.domain.qna.entity.Qna;
+import com.strawberryfarm.fitingle.domain.users.entity.Users;
+import com.strawberryfarm.fitingle.domain.wish.entity.Wish;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -22,7 +26,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
