--- src/main/java/com/strawberryfarm/fitingle/domain/Image.java
@@ -1,33 +1,0 @@
-package com.strawberryfarm.fitingle.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import lombok.Getter;
-
-@Entity
-@Getter
-@Table(name = "image")
-public class Image extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "boardId")
-    private Board board;
-
-    @Column(nullable = false)
-    private String imageUrl;
-
-    public void setBoard(Board board) {
-        this.board = board;
-    }
-}
