+++ src/main/java/com/strawberryfarm/fitingle/domain/image/entity/Image.java
@@ -1,5 +1,7 @@
-package com.strawberryfarm.fitingle.domain;
+package com.strawberryfarm.fitingle.domain.image.entity;
 
+import com.strawberryfarm.fitingle.domain.BaseEntity;
+import com.strawberryfarm.fitingle.domain.Board;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -7,15 +9,15 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
-import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import lombok.Getter;
 
 @Entity
 @Getter
-@Table(name = "tag")
-public class Tag extends BaseEntity {
+@Table(name = "image")
+public class Image extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -24,7 +26,10 @@ public class Tag extends BaseEntity {
     @JoinColumn(name = "boardId")
     private Board board;
 
-    @Lob
     @Column(nullable = false)
-    private String contents;
+    private String imageUrl;
+
+    public void setBoard(Board board) {
+        this.board = board;
+    }
 }
