+++ src/main/java/com/strawberryfarm/fitingle/domain/qna/entity/Qna.java
@@ -1,10 +1,10 @@
-package com.strawberryfarm.fitingle.domain;
+package com.strawberryfarm.fitingle.domain.qna.entity;
 
+import com.strawberryfarm.fitingle.domain.BaseEntity;
+import com.strawberryfarm.fitingle.domain.Board;
 import com.strawberryfarm.fitingle.domain.users.entity.Users;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -13,43 +13,38 @@ import javax.persistence.JoinColumn;
 import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@Table(name = "apply")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Apply extends BaseEntity{
-
+@Table(name = "qna")
+public class Qna extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
+    @JoinColumn(name = "userId")
     private Users user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "boardId", insertable = false, updatable = false)
+    private Board board;
+
 
     @Column(nullable = false)
     private Long boardId;
 
     @Lob
+    @Column(nullable = false)
     private String contents;
 
     @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private ApplyStatus status;
+    private boolean status;
+
+    private String password;
 
     public void setUser(Users users) {
-        this.user = users;
+        this.user =  users;
     }
-
-//    @Column(nullable = false)
-//    private LocalDateTime created_date;
-//
-//    @Column(nullable = false)
-//    private LocalDateTime updated_date;
 }
-
-
