+++ src/main/java/com/strawberryfarm/fitingle/domain/searchlog/entity/SearchLog.java
@@ -1,5 +1,6 @@
-package com.strawberryfarm.fitingle.domain;
+package com.strawberryfarm.fitingle.domain.searchlog.entity;
 
+import com.strawberryfarm.fitingle.domain.BaseEntity;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -10,15 +11,14 @@ import lombok.Getter;
 
 @Entity
 @Getter
-@Table(name = "keyword")
-public class Keyword extends BaseEntity{
+@Table(name = "search_log")
+public class SearchLog extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    private Long userId;
+    private String word;
 
-    @Column(nullable = false)
-    private String name;
 }
