+++ src/main/java/com/strawberryfarm/fitingle/domain/tag/entity/Tag.java
@@ -1,5 +1,7 @@
-package com.strawberryfarm.fitingle.domain;
+package com.strawberryfarm.fitingle.domain.tag.entity;
 
+import com.strawberryfarm.fitingle.domain.BaseEntity;
+import com.strawberryfarm.fitingle.domain.Board;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -7,27 +9,24 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
+import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import lombok.Getter;
 
 @Entity
 @Getter
-@Table(name = "interest_field")
-public class InterestField extends BaseEntity{
+@Table(name = "tag")
+public class Tag extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "fieldId")
-    private Field field;
+    @JoinColumn(name = "boardId")
+    private Board board;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "userId")
-    private Users user;
-
-    public void setUser(Users users) {
-        this.user = users;
-    }
+    @Lob
+    @Column(nullable = false)
+    private String contents;
 }
