+++ src/main/java/com/strawberryfarm/fitingle/domain/users/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.strawberryfarm.fitingle.repository;
+package com.strawberryfarm.fitingle.domain.users.repository;
 
-import com.strawberryfarm.fitingle.domain.Users;
+import com.strawberryfarm.fitingle.domain.users.entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<Users, Long> {
