+++ src/main/java/com/strawberryfarm/fitingle/domain/wish/entity/Wish.java
@@ -1,21 +1,23 @@
-package com.strawberryfarm.fitingle.domain;
+package com.strawberryfarm.fitingle.domain.wish.entity;
 
-import javax.persistence.Column;
+import com.strawberryfarm.fitingle.domain.BaseEntity;
+import com.strawberryfarm.fitingle.domain.Board;
+import com.strawberryfarm.fitingle.domain.users.entity.Users;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
-import javax.persistence.Lob;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import lombok.Getter;
 
 @Entity
 @Getter
-@Table(name = "qna")
-public class Qna extends BaseEntity {
+@Table(name = "wish")
+public class Wish extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -25,23 +27,10 @@ public class Qna extends BaseEntity {
     private Users user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "boardId", insertable = false, updatable = false)
+    @JoinColumn(name = "boardId")
     private Board board;
 
-
-    @Column(nullable = false)
-    private Long boardId;
-
-    @Lob
-    @Column(nullable = false)
-    private String contents;
-
-    @Column(nullable = false)
-    private boolean status;
-
-    private String password;
-
     public void setUser(Users users) {
-        this.user =  users;
+        this.user = users;
     }
 }
