+++ src/main/resources/dummyfile.txt
@@ -1,1 +1,2 @@
-더미파일d
+도커 이미지는 매번 갱신해줘야한다. 안그러면 이전 이미지가 사용되서 컨테이너가 만들어진다 씨이이바아알
+resource 파일이 임의로 만들어도 되는지 알수가 없다.
