+++ src/test/java/com/strawberryfarm/fitingle/service/TestServiceTest.java
@@ -6,8 +6,8 @@ import com.strawberryfarm.fitingle.domain.Board;
 import com.strawberryfarm.fitingle.domain.Days;
 import com.strawberryfarm.fitingle.domain.PostStatus;
 import com.strawberryfarm.fitingle.domain.Times;
-import com.strawberryfarm.fitingle.domain.Users;
-import com.strawberryfarm.fitingle.repository.UserRepository;
+import com.strawberryfarm.fitingle.domain.users.entity.Users;
+import com.strawberryfarm.fitingle.domain.users.repository.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
